src/requirements.txt
plotly==5.0.0
tqdm==4.66.2
dj-database-url==0.5.0
Django==2.2.13
django-allauth==0.41.0
django-anymail==7.0.0
django-environ==0.4.5
django-extensions==2.2.8
django-heroku==0.3.1
django-pandas==0.6.1
djangorestframework==3.12.2
pandas==1.0.2
channels==3.0.3
channels-redis==3.2.0
numpy==1.18.1
